fixup! apply debug server stash
