+++ demos/node/.vscode/launch.json
@@ -5,12 +5,6 @@
     "version": "0.2.0",
     "configurations": [
         {
-            "type": "node",
-            "request": "launch",
-            "name": "Launch Program",
-            "program": "${workspaceFolder}/main.js"
-        },
-        {
             "type": "pwa-node",
             "request": "launch",
             "name": "PWA: launch program",
